updated development docs and slight refactoring done
