--- .github/ISSUES/10-order-management.md
@@ -1,162 +1,0 @@
-# Order Management
-
-## Requirements
-### User Story
-"As a user/admin, I want to manage and track orders throughout their lifecycle"
-
-### Acceptance Criteria
-- [ ] Users can:
-  - View their order history
-  - Track order status
-  - View order details
-  - Cancel orders (if eligible)
-  - Download invoices
-- [ ] Admins can:
-  - View all orders
-  - Update order status
-  - Process refunds
-  - Generate reports
-  - Manage shipping
-- [ ] System should:
-  - Send order status notifications
-  - Update inventory automatically
-  - Generate order analytics
-  - Handle order cancellations
-  - Maintain order history
-
-## Technical Design
-
-### Backend Changes
-#### Models
-- OrderStatus Model:
-  ```javascript
-  {
-    order: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'Order',
-      required: true
-    },
-    status: {
-      type: String,
-      enum: [
-        'pending',
-        'processing',
-        'shipped',
-        'delivered',
-        'cancelled',
-        'refunded'
-      ],
-      required: true
-    },
-    note: String,
-    updated_by: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'User',
-      required: true
-    },
-    timestamp: {
-      type: Date,
-      default: Date.now
-    }
-  }
-  ```
-
-#### Controllers
-- OrderController:
-  - getOrders
-    - Pagination
-    - Filtering
-    - Status tracking
-  - updateOrderStatus
-    - Status validation
-    - Inventory updates
-    - Notifications
-  - processRefund
-    - Payment integration
-    - Inventory adjustment
-  - generateInvoice
-    - PDF generation
-    - Email delivery
-  - getOrderAnalytics
-    - Sales reports
-    - Status metrics
-    - Revenue analysis
-
-#### Routes
-- GET /api/orders
-- GET /api/orders/:id
-- PATCH /api/orders/:id/status
-- POST /api/orders/:id/refund
-- GET /api/orders/:id/invoice
-- GET /api/orders/analytics
-- POST /api/orders/:id/cancel
-
-#### Middleware
-- validateOrderStatus
-- checkOrderPermissions
-- handleInventoryUpdate
-- validateRefundEligibility
-- ensureActiveOrder
-
-### Frontend Changes
-#### Components
-- OrderList
-  - Filter controls
-  - Status indicators
-  - Search functionality
-- OrderDetail
-  - Status timeline
-  - Action buttons
-  - Invoice download
-- OrderAnalytics
-  - Status charts
-  - Revenue graphs
-  - Trend analysis
-- RefundManager
-  - Refund form
-  - Payment details
-  - Confirmation dialog
-
-#### Services
-- orderService:
-  - Order operations
-  - Status management
-  - Analytics processing
-  - Notification handling
-
-### Dependencies
-- Authentication System (#1)
-- Admin Dashboard (#3)
-- Product CRUD (#4)
-- Checkout Process (#9)
-- Payment Gateway Service
-- PDF Generation Service
-- Email Service
-
-### Testing Requirements
-- Unit Tests:
-  - Status transitions
-  - Refund calculations
-  - Permission checks
-- Integration Tests:
-  - Order flow
-  - Inventory updates
-  - Payment processing
-- E2E Tests:
-  - Order lifecycle
-  - Admin operations
-  - User interactions
-
-### Documentation Updates
-- API Documentation:
-  - Order endpoints
-  - Status workflows
-  - Analytics formats
-- User Guide:
-  - Order tracking
-  - Refund policy
-  - Status definitions
-- Admin Guide:
-  - Order processing
-  - Refund handling
-  - Analytics interpretation
