--- .github/ISSUES/101-email-verification-frontend.md
@@ -1,36 +1,0 @@
-# Email Verification Frontend Implementation
-
-## Requirements
-### User Story
-"As a user, I want to verify my email address after registration so I can access my account"
-
-### Acceptance Criteria
-- [ ] Show "Check your email" message after registration
-- [ ] Email verification page with:
-  - Loading state
-  - Success state
-  - Error state
-- [ ] Clear user feedback throughout the process
-
-## Technical Design
-
-### Frontend Changes
-#### Components
-- EmailVerification.jsx
-  - Handles token verification
-  - Shows verification status
-  - Provides next steps
-
-#### Pages
-- verify-email.page.jsx
-  - Routes to verification component
-  - Handles URL parameters
-
-### Dependencies
-- Authentication System (#1)
-- Email Service
-
-### Testing Requirements
-- [ ] Test registration flow
-- [ ] Test verification link
-- [ ] Test error states
