+++ .github/ISSUES/11-document-upload-system.md
@@ -1,0 +1,179 @@
+# Document Upload System
+
+## Requirements
+### User Story
+"As a user, I want to securely upload and manage required documents for my visa applications, with clear feedback on upload status and validation requirements"
+
+### Acceptance Criteria
+- [ ] Users can upload documents with size up to 10MB
+- [ ] System validates file types (PDF, JPG, PNG)
+- [ ] Users receive clear feedback on upload progress and status
+- [ ] Users can replace/update uploaded documents
+- [ ] Users can view their uploaded documents
+- [ ] System securely stores uploaded files
+- [ ] Admin can access and verify uploaded documents
+- [ ] System generates secure URLs for document access
+- [ ] Documents are properly associated with visa applications
+
+## Technical Design
+
+### Backend Changes
+#### Models
+- Document Model: ```javascript
+  {
+    userId: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'User',
+      required: true
+    },
+    applicationId: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'Application',
+      required: true
+    },
+    type: {
+      type: String,
+      enum: ['passport', 'photo', 'ticket', 'hotel', 'other'],
+      required: true
+    },
+    fileDetails: {
+      originalName: String,
+      mimeType: String,
+      size: Number,
+      extension: String
+    },
+    storage: {
+      provider: String,
+      bucket: String,
+      key: String,
+      url: String,
+      expiryDate: Date
+    },
+    status: {
+      type: String,
+      enum: ['pending', 'verified', 'rejected'],
+      default: 'pending'
+    },
+    verificationDetails: {
+      verifiedBy: {
+        type: mongoose.Schema.Types.ObjectId,
+        ref: 'User'
+      },
+      verifiedAt: Date,
+      notes: String
+    },
+    uploadedAt: {
+      type: Date,
+      default: Date.now
+    },
+    lastUpdated: Date,
+    isActive: {
+      type: Boolean,
+      default: true
+    }
+  } ```
+
+#### Controllers
+- DocumentController:
+  - uploadDocument
+  - validateDocument
+  - getDocument
+  - updateDocument
+  - deleteDocument
+  - verifyDocument
+  - generateSecureUrl
+
+#### Services
+- StorageService:
+  - initializeStorage
+  - uploadFile
+  - deleteFile
+  - generatePresignedUrl
+  - validateFileType
+  - validateFileSize
+  - scanForMalware
+
+#### Routes
+- POST /api/documents/upload
+- GET /api/documents/:id
+- PUT /api/documents/:id
+- DELETE /api/documents/:id
+- POST /api/documents/:id/verify
+- GET /api/documents/:id/download
+
+### Frontend Changes
+#### Components
+- DocumentUploader
+  - FileInput
+  - ProgressBar
+  - ValidationMessages
+  - FilePreview
+  - UploadStatus
+- DocumentManager
+  - DocumentList
+  - DocumentCard
+  - DocumentActions
+  - DocumentPreview
+- AdminDocumentVerification
+  - VerificationQueue
+  - DocumentReview
+  - VerificationActions
+
+#### Services
+- documentService:
+  - uploadHandling
+  - progressTracking
+  - statusManagement
+  - documentRetrieval
+  - errorHandling
+
+### Dependencies
+- Authentication System ✅ (#1)
+- User Profile Management ✅ (#2)
+- Storage Management (#13)
+
+### Special Features
+1. Security Measures
+   - Virus scanning before storage
+   - File type validation
+   - Secure URL generation
+   - Access control
+
+2. Performance Optimizations
+   - Client-side file compression
+   - Chunked uploads for large files
+   - Progressive loading for previews
+   - Caching of frequently accessed documents
+
+3. User Experience
+   - Drag and drop support
+   - Multiple file upload
+   - Upload progress tracking
+   - Immediate validation feedback
+   - Preview generation for supported file types
+
+4. Admin Features
+   - Batch document verification
+   - Document status management
+   - Audit trail for verifications
+   - Secure document viewing
+
+### Implementation Notes
+1. Security Requirements
+   - Implement virus scanning
+   - Enforce file type restrictions
+   - Set up secure storage
+   - Configure access controls
+   - Enable audit logging
+
+2. Storage Configuration
+   - Set up S3 or equivalent
+   - Configure backup system
+   - Implement file lifecycle
+   - Set up CDN for delivery
+
+3. Performance Considerations
+   - Implement upload size limits
+   - Configure timeout settings
+   - Set up error recovery
+   - Enable resumable uploads
