--- .github/ISSUES/11-stripe-integration.md
@@ -1,151 +1,0 @@
-# Stripe Integration
-
-## Requirements
-### User Story
-"As an admin, I want to integrate Stripe payment processing so that customers can securely pay for their orders"
-
-### Acceptance Criteria
-- [ ] Secure Stripe integration setup
-- [ ] Support for:
-  - Credit/debit cards
-  - Multiple currencies
-  - Automated payouts
-  - Refund processing
-- [ ] PCI compliance
-- [ ] Webhook handling for:
-  - Payment success/failure
-  - Refund status
-  - Dispute management
-- [ ] Error handling and logging
-- [ ] Test mode support
-
-## Technical Design
-
-### Backend Changes
-#### Models
-- PaymentConfig Model:
-  ```javascript
-  {
-    environment: {
-      type: String,
-      enum: ['test', 'production'],
-      required: true
-    },
-    stripe_public_key: {
-      type: String,
-      required: true
-    },
-    stripe_webhook_secret: {
-      type: String,
-      required: true
-    },
-    supported_currencies: [{
-      type: String,
-      default: ['USD']
-    }],
-    payment_methods: [{
-      type: String,
-      enum: ['card', 'ideal', 'sepa'],
-      default: ['card']
-    }],
-    created_by: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'User',
-      required: true
-    },
-    updated_at: {
-      type: Date,
-      default: Date.now
-    }
-  }
-  ```
-
-#### Services
-- stripeService:
-  - initializeStripe
-    - Environment setup
-    - API key configuration
-    - Webhook registration
-  - paymentIntents
-    - Create/confirm intents
-    - Handle 3D Secure
-    - Process refunds
-  - webhookHandler
-    - Event verification
-    - Payment updates
-    - Error handling
-  - customerManagement
-    - Create/update customers
-    - Save payment methods
-    - Handle subscriptions
-
-#### Routes
-- POST /api/stripe/config
-- POST /api/stripe/webhook
-- POST /api/payments/intent
-- POST /api/payments/confirm
-- POST /api/payments/refund
-- GET /api/payments/methods
-
-#### Middleware
-- validateStripeSignature
-- validatePaymentIntent
-- handleStripeErrors
-- checkStripeConfig
-- logStripeEvents
-
-### Frontend Changes
-#### Components
-- StripeProvider
-  - Context setup
-  - Elements wrapper
-  - Error boundary
-- PaymentForm
-  - Card element
-  - Payment request button
-  - Error handling
-- PaymentStatus
-  - Processing state
-  - Success/failure display
-  - Receipt generation
-
-#### Services
-- stripeClientService:
-  - Element initialization
-  - Payment processing
-  - Error handling
-  - Status management
-
-### Dependencies
-- Authentication System (#1)
-- Admin Dashboard (#3)
-- Checkout Process (#9)
-- Order Management (#10)
-
-### Testing Requirements
-- Unit Tests:
-  - Payment calculations
-  - Webhook validation
-  - Error handling
-- Integration Tests:
-  - Payment flow
-  - Refund process
-  - Webhook handling
-- E2E Tests:
-  - Complete payment flow
-  - Error scenarios
-  - Refund process
-
-### Documentation Updates
-- API Documentation:
-  - Stripe integration
-  - Webhook setup
-  - Error codes
-- Admin Guide:
-  - Stripe configuration
-  - Payment monitoring
-  - Dispute handling
-- Security Documentation:
-  - PCI compliance
-  - Key management
-  - Data handling
