+++ .github/ISSUES/12-document-verification.md
@@ -1,0 +1,196 @@
+# Document Verification
+
+## Requirements
+### User Story
+"As an admin, I want to efficiently verify and process uploaded documents for visa applications, ensuring all submissions meet the required standards and maintaining a clear audit trail"
+
+### Acceptance Criteria
+- [ ] Admin can view a queue of pending document verifications
+- [ ] Admin can examine documents in detail with necessary tools
+- [ ] Admin can approve or reject documents with comments
+- [ ] Admin can batch process similar documents
+- [ ] System maintains complete verification history
+- [ ] System notifies users of verification results
+- [ ] Admin can flag suspicious documents for review
+- [ ] System tracks verification metrics and performance
+- [ ] Admin can search and filter documents by status
+
+## Technical Design
+
+### Backend Changes
+#### Models
+- DocumentVerification Model: ```javascript
+  {
+    documentId: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'Document',
+      required: true
+    },
+    verifiedBy: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'User',
+      required: true
+    },
+    applicationId: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'Application',
+      required: true
+    },
+    status: {
+      type: String,
+      enum: ['pending', 'approved', 'rejected', 'flagged'],
+      default: 'pending'
+    },
+    verificationDetails: {
+      timestamp: Date,
+      decision: String,
+      comments: String,
+      checklistCompleted: Boolean
+    },
+    qualityMetrics: {
+      imageQuality: String,
+      readability: String,
+      completeness: String
+    },
+    securityChecks: {
+      authenticityVerified: Boolean,
+      malwareScanned: Boolean,
+      digitalSignatureValid: Boolean
+    },
+    auditTrail: [{
+      action: String,
+      performedBy: {
+        type: mongoose.Schema.Types.ObjectId,
+        ref: 'User'
+      },
+      timestamp: Date,
+      details: String
+    }],
+    processingTime: Number,
+    priority: {
+      type: String,
+      enum: ['low', 'medium', 'high', 'urgent'],
+      default: 'medium'
+    },
+    notificationsSent: [{
+      type: String,
+      timestamp: Date,
+      recipient: String,
+      status: String
+    }],
+    lastUpdated: Date
+  } ```
+
+#### Controllers
+- VerificationController:
+  - getVerificationQueue
+  - processVerification
+  - updateVerificationStatus
+  - addVerificationComment
+  - getVerificationHistory
+  - generateVerificationReport
+  - batchProcessVerifications
+  - flagForReview
+
+#### Services
+- VerificationService:
+  - documentQualityCheck
+  - authenticityVerification
+  - updateVerificationStatus
+  - notifyVerificationResult
+  - trackVerificationMetrics
+  - generateAuditTrail
+
+#### Routes
+- GET /api/verifications/queue
+- GET /api/verifications/:id
+- POST /api/verifications/:id/process
+- PUT /api/verifications/:id/status
+- POST /api/verifications/batch
+- GET /api/verifications/metrics
+- GET /api/verifications/history
+- POST /api/verifications/:id/flag
+
+### Frontend Changes
+#### Components
+- VerificationDashboard
+  - QueueOverview
+  - DocumentPreview
+  - VerificationTools
+  - DecisionPanel
+  - MetricsDisplay
+- VerificationWorkflow
+  - DocumentChecklist
+  - QualityAssessment
+  - SecurityVerification
+  - DecisionForm
+  - BatchProcessor
+- VerificationHistory
+  - AuditTrail
+  - TimelineView
+  - SearchFilters
+  - ReportGenerator
+
+#### Services
+- verificationService:
+  - queueManagement
+  - documentProcessing
+  - metricsTracking
+  - notificationHandling
+  - reportGeneration
+
+### Dependencies
+- Authentication System ✅ (#1)
+- Document Upload System ✅ (#11)
+- Storage Management (#13)
+
+### Special Features
+1. Verification Tools
+   - Image enhancement tools
+   - Document comparison
+   - Digital signature verification
+   - Automated quality checks
+   - OCR for text verification
+
+2. Workflow Optimization
+   - Priority queue management
+   - Batch processing capabilities
+   - Automated pre-checks
+   - Performance metrics tracking
+   - Workload distribution
+
+3. Security Features
+   - Audit logging
+   - Access control
+   - Digital signature validation
+   - Suspicious activity detection
+   - Verification history tracking
+
+4. Admin Experience
+   - Customizable verification checklists
+   - Quick decision shortcuts
+   - Batch processing templates
+   - Performance dashboards
+   - Document comparison tools
+
+### Implementation Notes
+1. Verification Process
+   - Define verification criteria
+   - Implement quality checks
+   - Set up notification system
+   - Configure audit logging
+   - Establish verification SLAs
+
+2. Performance Requirements
+   - Queue optimization
+   - Response time targets
+   - Batch processing limits
+   - Resource allocation
+   - Caching strategies
+
+3. Security Considerations
+   - Access control matrix
+   - Audit trail requirements
+   - Data retention policies
+   - Compliance requirements
+   - Security logging
