--- .github/ISSUES/12-payment-processing.md
@@ -1,174 +1,0 @@
-# Payment Processing
-
-## Requirements
-### User Story
-"As a customer, I want to securely process my payment so that I can complete my purchase"
-
-### Acceptance Criteria
-- [ ] Secure payment processing with Stripe
-- [ ] Support for:
-  - Credit/debit cards
-  - Multiple currencies
-  - 3D Secure authentication
-  - Payment failure handling
-- [ ] Real-time payment status updates
-- [ ] Payment receipt generation
-- [ ] Transaction history
-- [ ] Refund processing capability
-- [ ] Error handling and logging
-
-## Technical Design
-
-### Backend Changes
-#### Models
-- Transaction Model:
-  ```javascript
-  {
-    order: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'Order',
-      required: true
-    },
-    stripe_payment_intent_id: {
-      type: String,
-      required: true,
-      unique: true
-    },
-    amount: {
-      type: Number,
-      required: true
-    },
-    currency: {
-      type: String,
-      required: true,
-      default: 'USD'
-    },
-    status: {
-      type: String,
-      enum: [
-        'pending',
-        'processing',
-        'succeeded',
-        'failed',
-        'refunded',
-        'partially_refunded'
-      ],
-      default: 'pending'
-    },
-    payment_method: {
-      type: String,
-      required: true
-    },
-    metadata: {
-      customer_email: String,
-      shipping_address: Object,
-      items: Array
-    },
-    error_log: [{
-      code: String,
-      message: String,
-      timestamp: {
-        type: Date,
-        default: Date.now
-      }
-    }],
-    created_at: {
-      type: Date,
-      default: Date.now
-    }
-  }
-  ```
-
-#### Services
-- paymentService:
-  - processPayment
-    - Create payment intent
-    - Handle 3D Secure
-    - Process confirmation
-  - handleRefund
-    - Full/partial refund
-    - Inventory updates
-    - Email notifications
-  - paymentWebhooks
-    - Status updates
-    - Error handling
-    - Order updates
-  - transactionHistory
-    - Payment records
-    - Refund tracking
-    - Analytics data
-
-#### Routes
-- POST /api/payments/process
-- POST /api/payments/confirm
-- POST /api/payments/refund
-- GET /api/payments/status/:id
-- GET /api/payments/history
-- POST /api/payments/webhook
-
-#### Middleware
-- validatePaymentData
-- handleStripeErrors
-- verifyWebhookSignature
-- logTransactions
-- checkRefundEligibility
-
-### Frontend Changes
-#### Components
-- PaymentProcessor
-  - Payment form
-  - 3D Secure handling
-  - Status updates
-- PaymentStatus
-  - Processing indicator
-  - Success/failure display
-  - Error messages
-- TransactionHistory
-  - Payment list
-  - Refund requests
-  - Receipt access
-- RefundInterface
-  - Refund form
-  - Status tracking
-  - Confirmation dialog
-
-#### Services
-- paymentClientService:
-  - Payment processing
-  - Status checking
-  - Receipt generation
-  - Error handling
-
-### Dependencies
-- Authentication System (#1)
-- Checkout Process (#9)
-- Order Management (#10)
-- Stripe Integration (#11)
-
-### Testing Requirements
-- Unit Tests:
-  - Payment validation
-  - Refund calculations
-  - Error handling
-- Integration Tests:
-  - Payment flow
-  - Webhook handling
-  - Refund process
-- E2E Tests:
-  - Complete payment flow
-  - 3D Secure handling
-  - Error scenarios
-
-### Documentation Updates
-- API Documentation:
-  - Payment endpoints
-  - Webhook events
-  - Error codes
-- User Guide:
-  - Payment methods
-  - Refund policy
-  - Transaction history
-- Security Documentation:
-  - PCI compliance
-  - Data handling
-  - Error logging
