--- .github/ISSUES/13-order-confirmation.md
@@ -1,151 +1,0 @@
-# Order Confirmation
-
-## Requirements
-### User Story
-"As a customer, I want to receive confirmation of my order so that I know my purchase was successful"
-
-### Acceptance Criteria
-- [ ] Users receive:
-  - Order confirmation page
-  - Email confirmation
-  - Order tracking number
-  - Purchase receipt
-- [ ] Confirmation includes:
-  - Order details
-  - Shipping information
-  - Payment summary
-  - Estimated delivery
-  - Next steps
-- [ ] Support for:
-  - Guest orders
-  - Registered users
-  - Multiple email formats
-  - Order tracking links
-
-## Technical Design
-
-### Backend Changes
-#### Models
-- OrderConfirmation Model:
-  ```javascript
-  {
-    order: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'Order',
-      required: true,
-      unique: true
-    },
-    confirmation_number: {
-      type: String,
-      required: true,
-      unique: true
-    },
-    email_status: {
-      type: String,
-      enum: ['pending', 'sent', 'failed', 'delivered'],
-      default: 'pending'
-    },
-    tracking_number: String,
-    email_attempts: [{
-      timestamp: Date,
-      status: String,
-      error: String
-    }],
-    receipt_url: String,
-    created_at: {
-      type: Date,
-      default: Date.now
-    }
-  }
-  ```
-
-#### Services
-- confirmationService:
-  - generateConfirmation
-    - Create confirmation record
-    - Generate unique numbers
-    - Create receipt PDF
-  - sendConfirmationEmail
-    - Email template processing
-    - Attachment handling
-    - Retry logic
-  - trackingService
-    - Generate tracking number
-    - Update shipping status
-    - Delivery estimates
-
-#### Routes
-- POST /api/orders/:id/confirm
-- GET /api/orders/:id/confirmation
-- GET /api/orders/:id/receipt
-- POST /api/orders/:id/resend-confirmation
-- GET /api/orders/:id/tracking
-
-#### Middleware
-- validateOrderStatus
-- generateConfirmationNumber
-- prepareConfirmationData
-- handleEmailDelivery
-
-### Frontend Changes
-#### Components
-- OrderConfirmationPage
-  - Success message
-  - Order summary
-  - Tracking info
-  - Next steps
-- ConfirmationEmail
-  - Email template
-  - Mobile responsive
-  - PDF receipt
-- TrackingDisplay
-  - Status timeline
-  - Delivery estimate
-  - Shipping updates
-- ReceiptViewer
-  - PDF display
-  - Download option
-  - Print layout
-
-#### Services
-- confirmationClientService:
-  - Confirmation display
-  - Receipt handling
-  - Email resend
-  - Tracking updates
-
-### Dependencies
-- Authentication System (#1)
-- Order Management (#10)
-- Payment Processing (#12)
-- Email Service
-- PDF Generation Service
-- Shipping Tracking Service
-
-### Testing Requirements
-- Unit Tests:
-  - Confirmation generation
-  - Email formatting
-  - Receipt creation
-- Integration Tests:
-  - Email delivery
-  - PDF generation
-  - Tracking updates
-- E2E Tests:
-  - Complete confirmation flow
-  - Email receipt
-  - Tracking display
-
-### Documentation Updates
-- API Documentation:
-  - Confirmation endpoints
-  - Email templates
-  - Receipt format
-- User Guide:
-  - Order tracking
-  - Email expectations
-  - Receipt access
-- System Guide:
-  - Email configuration
-  - PDF generation
-  - Tracking integration
