+++ .github/ISSUES/13-storage-management.md
@@ -1,0 +1,212 @@
+# Storage Management
+
+## Requirements
+### User Story
+"As a system administrator, I want to efficiently manage and secure document storage, ensuring proper lifecycle management, backup procedures, and quick retrieval while maintaining compliance with data protection regulations"
+
+### Acceptance Criteria
+- [ ] System securely stores all uploaded documents
+- [ ] Documents are automatically organized in logical structures
+- [ ] System implements proper backup and recovery procedures
+- [ ] Storage capacity is monitored and managed efficiently
+- [ ] Access to documents is properly controlled and logged
+- [ ] System implements document lifecycle management
+- [ ] Storage costs are optimized through tiered storage
+- [ ] System maintains data retention compliance
+- [ ] Quick document retrieval is supported
+
+## Technical Design
+
+### Backend Changes
+#### Models
+- StorageManagement Model: ```javascript
+  {
+    resourceId: {
+      type: mongoose.Schema.Types.ObjectId,
+      required: true
+    },
+    resourceType: {
+      type: String,
+      enum: ['document', 'application', 'user'],
+      required: true
+    },
+    storageDetails: {
+      provider: {
+        type: String,
+        enum: ['s3', 'azure', 'local'],
+        required: true
+      },
+      bucket: String,
+      path: String,
+      size: Number,
+      lastAccessed: Date,
+      accessCount: Number
+    },
+    lifecycle: {
+      createdAt: Date,
+      expiresAt: Date,
+      retentionPeriod: Number,
+      archiveDate: Date,
+      deletionDate: Date
+    },
+    encryption: {
+      algorithm: String,
+      keyId: String,
+      status: String
+    },
+    backupStatus: {
+      lastBackup: Date,
+      backupLocation: String,
+      verified: Boolean
+    },
+    accessControl: {
+      permissions: [{
+        role: String,
+        access: String
+      }],
+      publicAccessible: Boolean,
+      temporaryUrls: [{
+        url: String,
+        expiresAt: Date
+      }]
+    },
+    metrics: {
+      storageClass: String,
+      monthlyAccessCount: Number,
+      monthlyBandwidth: Number,
+      monthlyCost: Number
+    },
+    status: {
+      type: String,
+      enum: ['active', 'archived', 'deleted'],
+      default: 'active'
+    },
+    tags: [{
+      key: String,
+      value: String
+    }]
+  } ```
+
+#### Controllers
+- StorageController:
+  - initializeStorage
+  - storeResource
+  - retrieveResource
+  - deleteResource
+  - updateResourceMetadata
+  - generatePresignedUrl
+  - manageLifecycle
+  - monitorUsage
+  - manageBackups
+
+#### Services
+- StorageService:
+  - resourceUpload
+  - resourceDownload
+  - backupManagement
+  - lifecycleManagement
+  - encryptionHandling
+  - accessControl
+  - usageMetrics
+  - costOptimization
+
+#### Routes
+- POST /api/storage/resources
+- GET /api/storage/resources/:id
+- PUT /api/storage/resources/:id
+- DELETE /api/storage/resources/:id
+- GET /api/storage/metrics
+- POST /api/storage/backup
+- GET /api/storage/usage
+- POST /api/storage/cleanup
+
+### Frontend Changes
+#### Components
+- StorageAdminDashboard
+  - UsageMetrics
+  - StorageMonitor
+  - BackupStatus
+  - ResourceBrowser
+  - CostAnalytics
+- StorageManagement
+  - ResourceManager
+  - LifecycleConfig
+  - BackupManager
+  - AccessControl
+  - MetricsViewer
+- StorageAnalytics
+  - UsageGraphs
+  - CostReports
+  - AccessPatterns
+  - PerformanceMetrics
+
+#### Services
+- storageService:
+  - resourceManagement
+  - backupOperations
+  - metricsCollection
+  - accessManagement
+  - lifecycleHandling
+
+### Dependencies
+- Authentication System ✅ (#1)
+- Document Upload System ✅ (#11)
+- Document Verification ✅ (#12)
+
+### Special Features
+1. Storage Optimization
+   - Automatic tiering
+   - Compression algorithms
+   - Deduplication
+   - Access-based optimization
+   - Cost-effective storage classes
+
+2. Security Measures
+   - Encryption at rest
+   - Encryption in transit
+   - Key rotation
+   - Access logging
+   - Compliance monitoring
+
+3. Backup Features
+   - Automated backups
+   - Point-in-time recovery
+   - Cross-region replication
+   - Backup verification
+   - Disaster recovery
+
+4. Performance Features
+   - Caching layer
+   - CDN integration
+   - Quick retrieval
+   - Bulk operations
+   - Performance monitoring
+
+### Implementation Notes
+1. Storage Configuration
+   - Set up storage providers
+   - Configure backup systems
+   - Implement monitoring
+   - Set up alerts
+   - Define storage classes
+
+2. Security Setup
+   - Configure encryption
+   - Set up access controls
+   - Implement audit logging
+   - Define security policies
+   - Set up compliance monitoring
+
+3. Performance Optimization
+   - Configure caching
+   - Set up CDN
+   - Optimize retrieval
+   - Monitor metrics
+   - Implement scaling
+
+4. Cost Management
+   - Define storage tiers
+   - Set up cost monitoring
+   - Implement optimization rules
+   - Configure alerts
+   - Track usage patterns
