--- .github/ISSUES/14-email-notification-system.md
@@ -1,195 +1,0 @@
-# Email Notification System
-
-## Requirements
-### User Story
-"As a system administrator, I want to manage automated email notifications so that users receive timely and relevant communications"
-
-### Acceptance Criteria
-- [ ] System supports:
-  - Transactional emails
-    - Account verification
-    - Password reset
-    - Order confirmation
-    - Shipping updates
-    - Payment receipts
-  - Template management
-    - Custom templates
-    - Dynamic content
-    - Responsive design
-  - Email tracking
-    - Delivery status
-    - Open rates
-    - Click tracking
-- [ ] Admin features:
-  - Template management
-  - Email logs
-  - Delivery reports
-- [ ] Error handling:
-  - Retry mechanism
-  - Failure notifications
-  - Bounce handling
-
-## Technical Design
-
-### Backend Changes
-#### Models
-- EmailTemplate Model:
-  ```javascript
-  {
-    name: {
-      type: String,
-      required: true,
-      unique: true
-    },
-    type: {
-      type: String,
-      enum: [
-        'account_verification',
-        'password_reset',
-        'order_confirmation',
-        'shipping_update',
-        'payment_receipt'
-      ],
-      required: true
-    },
-    subject: {
-      type: String,
-      required: true
-    },
-    html_content: {
-      type: String,
-      required: true
-    },
-    text_content: String,
-    variables: [{
-      name: String,
-      description: String
-    }],
-    active: {
-      type: Boolean,
-      default: true
-    },
-    created_by: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'User',
-      required: true
-    },
-    updated_at: Date
-  }
-  ```
-
-- EmailLog Model:
-  ```javascript
-  {
-    template: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'EmailTemplate',
-      required: true
-    },
-    recipient: {
-      email: String,
-      user: {
-        type: mongoose.Schema.Types.ObjectId,
-        ref: 'User'
-      }
-    },
-    status: {
-      type: String,
-      enum: ['queued', 'sent', 'delivered', 'failed', 'bounced'],
-      default: 'queued'
-    },
-    metadata: {
-      order_id: String,
-      verification_token: String,
-      reset_token: String
-    },
-    tracking: {
-      opened: Boolean,
-      opened_at: Date,
-      clicked: Boolean,
-      clicked_at: Date
-    },
-    error: {
-      code: String,
-      message: String,
-      attempts: Number
-    }
-  }
-  ```
-
-#### Services
-- emailService:
-  - Template management
-  - Email sending
-  - Status tracking
-  - Error handling
-- queueService:
-  - Email queuing
-  - Retry logic
-  - Rate limiting
-
-#### Routes
-- POST /api/email/templates
-- GET /api/email/templates
-- PUT /api/email/templates/:id
-- DELETE /api/email/templates/:id
-- GET /api/email/logs
-- POST /api/email/test
-- GET /api/email/stats
-
-### Frontend Changes
-#### Components
-- EmailTemplateManager
-  - Template CRUD
-  - Variable management
-  - Preview function
-- EmailLogs
-  - Log viewer
-  - Status tracking
-  - Error details
-- EmailStats
-  - Delivery metrics
-  - Open rates
-  - Click rates
-
-#### Services
-- emailAdminService:
-  - Template management
-  - Log viewing
-  - Stats tracking
-
-### Dependencies
-- Authentication System (#1)
-- Admin Dashboard (#3)
-- Order Management (#10)
-- Order Confirmation (#13)
-- Email Service Provider (SendGrid/AWS SES)
-- Queue System (Redis/Bull)
-
-### Testing Requirements
-- Unit Tests:
-  - Template rendering
-  - Variable validation
-  - Queue processing
-- Integration Tests:
-  - Email sending
-  - Status tracking
-  - Error handling
-- E2E Tests:
-  - Template management
-  - Email delivery
-  - Tracking system
-
-### Documentation Updates
-- API Documentation:
-  - Email endpoints
-  - Template format
-  - Variable system
-- Admin Guide:
-  - Template management
-  - Email best practices
-  - Troubleshooting
-- Technical Guide:
-  - Email configuration
-  - Queue setup
-  - Error handling
