+++ .github/ISSUES/14-stripe-integration.md
@@ -1,0 +1,213 @@
+# Stripe Integration
+
+## Requirements
+### User Story
+"As a user, I want to securely pay for my visa application using a credit card, receiving immediate confirmation and being able to track my payment status"
+
+### Acceptance Criteria
+- [ ] Users can securely submit credit card payments
+- [ ] System processes payments through Stripe
+- [ ] Users receive immediate payment confirmation
+- [ ] System handles failed payments gracefully
+- [ ] Admin can view and manage payment transactions
+- [ ] System maintains payment history
+- [ ] Refunds can be processed when necessary
+- [ ] Payment webhooks are properly handled
+- [ ] System provides detailed payment receipts
+
+## Technical Design
+
+### Backend Changes
+#### Models
+- Payment Model: ```javascript
+  {
+    userId: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'User',
+      required: true
+    },
+    applicationId: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'Application',
+      required: true
+    },
+    stripeDetails: {
+      customerId: String,
+      paymentIntentId: String,
+      chargeId: String,
+      receiptUrl: String,
+      paymentMethodId: String
+    },
+    amount: {
+      value: Number,
+      currency: {
+        type: String,
+        default: 'EUR'
+      }
+    },
+    status: {
+      type: String,
+      enum: ['pending', 'processing', 'completed', 'failed', 'refunded'],
+      default: 'pending'
+    },
+    metadata: {
+      visaType: String,
+      applicationType: String,
+      processingSpeed: String
+    },
+    paymentMethod: {
+      type: String,
+      brand: String,
+      last4: String,
+      expiryMonth: Number,
+      expiryYear: Number
+    },
+    billingDetails: {
+      name: String,
+      email: String,
+      address: {
+        line1: String,
+        line2: String,
+        city: String,
+        country: String,
+        postalCode: String
+      }
+    },
+    timestamps: {
+      created: Date,
+      processed: Date,
+      updated: Date
+    },
+    refund: {
+      refundId: String,
+      amount: Number,
+      reason: String,
+      status: String,
+      date: Date
+    },
+    webhookEvents: [{
+      eventType: String,
+      timestamp: Date,
+      data: Object
+    }]
+  } ```
+
+#### Controllers
+- PaymentController:
+  - createPaymentIntent
+  - confirmPayment
+  - retrievePayment
+  - processRefund
+  - getPaymentHistory
+  - handleWebhook
+  - updatePaymentStatus
+  - generateReceipt
+
+#### Services
+- StripeService:
+  - initializeStripe
+  - createCustomer
+  - processPayment
+  - handleRefund
+  - manageSubscription
+  - validateWebhook
+  - generatePaymentReceipt
+  - trackPaymentMetrics
+
+#### Routes
+- POST /api/payments/intent
+- POST /api/payments/confirm
+- GET /api/payments/:id
+- POST /api/payments/:id/refund
+- GET /api/payments/history
+- POST /api/payments/webhook
+- GET /api/payments/receipt/:id
+
+### Frontend Changes
+#### Components
+- PaymentProcessor
+  - CardElement
+  - PaymentForm
+  - BillingDetails
+  - PaymentStatus
+  - ConfirmationDisplay
+- PaymentDashboard
+  - TransactionList
+  - PaymentDetails
+  - RefundInterface
+  - PaymentMetrics
+- PaymentHistory
+  - TransactionTimeline
+  - ReceiptViewer
+  - RefundStatus
+  - PaymentSearch
+
+#### Services
+- paymentService:
+  - stripeIntegration
+  - paymentProcessing
+  - refundHandling
+  - receiptGeneration
+  - statusTracking
+
+### Dependencies
+- Authentication System ✅ (#1)
+- User Profile Management ✅ (#2)
+
+### Special Features
+1. Payment Processing
+   - Real-time card validation
+   - 3D Secure support
+   - Multiple currency support
+   - Automatic retry logic
+   - Payment optimization
+
+2. Security Features
+   - PCI compliance
+   - Fraud detection
+   - Data encryption
+   - Secure token handling
+   - Error logging
+
+3. Admin Features
+   - Transaction monitoring
+   - Refund management
+   - Payment analytics
+   - Dispute handling
+   - Automated reconciliation
+
+4. User Experience
+   - Save payment methods
+   - Quick checkout
+   - Payment status tracking
+   - Email notifications
+   - Digital receipts
+
+### Implementation Notes
+1. Stripe Setup
+   - Configure API keys
+   - Set up webhooks
+   - Enable required features
+   - Configure payment methods
+   - Set up test environment
+
+2. Security Requirements
+   - PCI compliance
+   - Data encryption
+   - Error handling
+   - Logging setup
+   - Access control
+
+3. Testing Strategy
+   - Test card numbers
+   - Webhook testing
+   - Error scenarios
+   - Refund testing
+   - Integration testing
+
+4. Monitoring Setup
+   - Payment tracking
+   - Error monitoring
+   - Performance metrics
+   - Webhook reliability
+   - System health checks
