+++ .github/ISSUES/15-payment-processing.md
@@ -1,0 +1,216 @@
+# Payment Processing
+
+## Requirements
+### User Story
+"As a user, I want a smooth payment process for my visa application with clear pricing, flexible payment options, and transparent status updates throughout the transaction"
+
+### Acceptance Criteria
+- [ ] System displays clear pricing for different visa types
+- [ ] Users can select payment method (credit card, saved cards)
+- [ ] System validates payment information before processing
+- [ ] Users receive step-by-step payment status updates
+- [ ] System handles payment failures with clear error messages
+- [ ] Users receive payment confirmation and receipt
+- [ ] System updates application status after successful payment
+- [ ] Admin can manage pricing and payment settings
+- [ ] System maintains detailed payment records
+
+## Technical Design
+
+### Backend Changes
+#### Models
+- PaymentProcess Model: ```javascript
+  {
+    applicationId: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'Application',
+      required: true
+    },
+    pricing: {
+      basePrice: Number,
+      additionalFees: [{
+        type: String,
+        amount: Number,
+        description: String
+      }],
+      discounts: [{
+        type: String,
+        amount: Number,
+        code: String,
+        expiryDate: Date
+      }],
+      total: Number,
+      currency: {
+        type: String,
+        default: 'EUR'
+      }
+    },
+    paymentFlow: {
+      currentStep: {
+        type: String,
+        enum: ['initiated', 'processing', 'authorized', 'captured', 'completed', 'failed'],
+        default: 'initiated'
+      },
+      startedAt: Date,
+      completedAt: Date,
+      timeoutAt: Date
+    },
+    transactionDetails: {
+      paymentId: {
+        type: mongoose.Schema.Types.ObjectId,
+        ref: 'Payment'
+      },
+      attempts: [{
+        timestamp: Date,
+        status: String,
+        errorCode: String,
+        errorMessage: String
+      }],
+      successfulAttempt: {
+        timestamp: Date,
+        confirmationCode: String
+      }
+    },
+    notifications: [{
+      type: String,
+      status: String,
+      timestamp: Date,
+      recipient: String,
+      template: String
+    }],
+    businessRules: {
+      priceRuleApplied: String,
+      taxRuleApplied: String,
+      discountRuleApplied: String
+    },
+    auditTrail: [{
+      action: String,
+      timestamp: Date,
+      performedBy: {
+        type: mongoose.Schema.Types.ObjectId,
+        ref: 'User'
+      },
+      details: String
+    }]
+  } ```
+
+#### Controllers
+- PaymentProcessController:
+  - initializePayment
+  - calculatePrice
+  - applyDiscounts
+  - validatePayment
+  - processTransaction
+  - handleCompletion
+  - manageFailure
+  - generateReceipts
+
+#### Services
+- PaymentProcessService:
+  - priceCalculation
+  - discountManagement
+  - paymentValidation
+  - transactionProcessing
+  - notificationHandling
+  - receiptGeneration
+  - statusTracking
+  - errorHandling
+
+#### Routes
+- POST /api/payment-process/initialize
+- GET /api/payment-process/calculate-price
+- POST /api/payment-process/validate
+- POST /api/payment-process/execute
+- GET /api/payment-process/:id/status
+- POST /api/payment-process/complete
+- GET /api/payment-process/pricing
+- POST /api/payment-process/discounts
+
+### Frontend Changes
+#### Components
+- PaymentWorkflow
+  - PricingDisplay
+  - PaymentMethodSelector
+  - ValidationDisplay
+  - ProcessingStatus
+  - ConfirmationView
+- PricingManagement
+  - PriceCalculator
+  - DiscountManager
+  - TaxCalculator
+  - FeeDisplay
+- TransactionMonitor
+  - StatusTracker
+  - ErrorHandler
+  - NotificationCenter
+  - ReceiptGenerator
+
+#### Services
+- paymentProcessService:
+  - workflowManagement
+  - priceCalculation
+  - statusTracking
+  - notificationHandling
+  - errorManagement
+
+### Dependencies
+- Authentication System ✅ (#1)
+- Stripe Integration ✅ (#14)
+
+### Special Features
+1. Price Management
+   - Dynamic pricing rules
+   - Discount system
+   - Tax calculation
+   - Fee management
+   - Currency conversion
+
+2. Payment Flow
+   - Step-by-step tracking
+   - Automatic retries
+   - Timeout handling
+   - Error recovery
+   - Status notifications
+
+3. Business Rules
+   - Price calculation rules
+   - Discount validation
+   - Tax rules
+   - Payment restrictions
+   - Refund policies
+
+4. Reporting Features
+   - Transaction reports
+   - Revenue analytics
+   - Payment statistics
+   - Error tracking
+   - Audit logging
+
+### Implementation Notes
+1. Business Logic
+   - Define pricing rules
+   - Set up discount system
+   - Configure tax rules
+   - Implement validation logic
+   - Set up audit logging
+
+2. Process Flow
+   - Define payment steps
+   - Set up status tracking
+   - Configure notifications
+   - Implement error handling
+   - Set up completion rules
+
+3. Integration Points
+   - Connect with Stripe
+   - Link to applications
+   - Set up notifications
+   - Configure reporting
+   - Enable monitoring
+
+4. Testing Requirements
+   - Price calculations
+   - Payment flows
+   - Error scenarios
+   - Business rules
+   - Integration tests
