--- .github/ISSUES/16-analytics-dashboard.md
@@ -1,163 +1,0 @@
-# Analytics Dashboard
-
-## Requirements
-### User Story
-"As an admin, I want to view comprehensive analytics so that I can make data-driven business decisions"
-
-### Acceptance Criteria
-- [ ] Dashboard displays:
-  - Sales metrics
-    - Daily/weekly/monthly sales
-    - Average order value
-    - Revenue trends
-    - Conversion rates
-  - Product analytics
-    - Best/worst sellers
-    - Stock turnover
-    - Category performance
-    - Search trends
-  - User analytics
-    - New users
-    - Active users
-    - Purchase frequency
-    - Cart abandonment
-  - Performance metrics
-    - Page views
-    - Session duration
-    - Bounce rates
-    - Load times
-
-## Technical Design
-
-### Backend Changes
-#### Models
-- AnalyticsEvent Model:
-  ```javascript
-  {
-    type: {
-      type: String,
-      enum: [
-        'page_view',
-        'product_view',
-        'cart_add',
-        'cart_remove',
-        'checkout_start',
-        'checkout_complete',
-        'search',
-        'user_register',
-        'user_login'
-      ],
-      required: true,
-      index: true
-    },
-    user: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'User',
-      index: true
-    },
-    session_id: String,
-    metadata: {
-      url: String,
-      product_id: mongoose.Schema.Types.ObjectId,
-      category_id: mongoose.Schema.Types.ObjectId,
-      search_query: String,
-      device_type: String,
-      referrer: String
-    },
-    timestamp: {
-      type: Date,
-      default: Date.now,
-      index: true
-    }
-  }
-  ```
-
-#### Services
-- analyticsService:
-  - Event tracking
-  - Data aggregation
-  - Report generation
-  - Trend analysis
-- reportingService:
-  - Custom reports
-  - Data export
-  - Scheduled reports
-  - Alert thresholds
-
-#### Routes
-- GET /api/analytics/dashboard
-- GET /api/analytics/sales
-- GET /api/analytics/products
-- GET /api/analytics/users
-- GET /api/analytics/performance
-- POST /api/analytics/reports/generate
-- GET /api/analytics/reports/:id
-- GET /api/analytics/export
-
-### Frontend Changes
-#### Components
-- AnalyticsDashboard
-  - Overview widgets
-  - Chart displays
-  - Date range selector
-  - Export options
-- SalesAnalytics
-  - Revenue charts
-  - Order metrics
-  - Payment analysis
-- ProductAnalytics
-  - Product performance
-  - Category insights
-  - Search analytics
-- UserAnalytics
-  - User behavior
-  - Demographics
-  - Engagement metrics
-- PerformanceMetrics
-  - Site performance
-  - Error tracking
-  - Load times
-
-#### Services
-- analyticsClientService:
-  - Data fetching
-  - Chart rendering
-  - Export handling
-  - Real-time updates
-
-### Dependencies
-- Admin Dashboard (#3)
-- Product Management (#4-7)
-- Order Management (#10)
-- Payment Processing (#12)
-- Inventory Management (#15)
-- Analytics Service (Google/Custom)
-- Export Service (CSV/PDF)
-
-### Testing Requirements
-- Unit Tests:
-  - Data calculations
-  - Report generation
-  - Export formatting
-- Integration Tests:
-  - Data aggregation
-  - Chart rendering
-  - Export process
-- E2E Tests:
-  - Dashboard navigation
-  - Report creation
-  - Data visualization
-
-### Documentation Updates
-- API Documentation:
-  - Analytics endpoints
-  - Report formats
-  - Export options
-- Admin Guide:
-  - Dashboard usage
-  - Report creation
-  - Data interpretation
-- Technical Guide:
-  - Event tracking
-  - Data aggregation
-  - Performance optimization
