+++ .github/ISSUES/16-payment-confirmation.md
@@ -1,0 +1,216 @@
+# Payment Confirmation
+
+## Requirements
+### User Story
+"As a user, I want to receive immediate confirmation of my payment with detailed transaction information and next steps for my visa application"
+
+### Acceptance Criteria
+- [ ] Users receive immediate on-screen confirmation
+- [ ] System sends detailed confirmation email
+- [ ] Users can view and download payment receipt
+- [ ] System updates application status automatically
+- [ ] Users can access payment history and details
+- [ ] Admin receives payment notification
+- [ ] System maintains confirmation audit trail
+- [ ] Confirmation includes next steps information
+- [ ] Users can share/download confirmation details
+
+## Technical Design
+
+### Backend Changes
+#### Models
+- PaymentConfirmation Model: ```javascript
+  {
+    paymentId: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'Payment',
+      required: true
+    },
+    applicationId: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'Application',
+      required: true
+    },
+    confirmationDetails: {
+      confirmationNumber: {
+        type: String,
+        unique: true
+      },
+      timestamp: Date,
+      status: {
+        type: String,
+        enum: ['confirmed', 'pending_verification', 'failed'],
+        default: 'confirmed'
+      },
+      amount: {
+        value: Number,
+        currency: String
+      }
+    },
+    communicationStatus: {
+      onScreen: {
+        displayed: Boolean,
+        timestamp: Date
+      },
+      email: {
+        sent: Boolean,
+        timestamp: Date,
+        recipient: String,
+        messageId: String
+      },
+      adminNotification: {
+        sent: Boolean,
+        timestamp: Date,
+        recipient: String
+      }
+    },
+    receiptDetails: {
+      receiptNumber: String,
+      generatedAt: Date,
+      url: String,
+      downloadCount: Number,
+      lastDownloaded: Date
+    },
+    nextSteps: {
+      instructions: [String],
+      deadlines: [{
+        action: String,
+        date: Date
+      }],
+      additionalRequirements: [String]
+    },
+    transactionSummary: {
+      paymentMethod: String,
+      last4: String,
+      merchantReference: String,
+      processorResponse: String
+    },
+    auditTrail: [{
+      action: String,
+      timestamp: Date,
+      performedBy: String,
+      details: String
+    }]
+  } ```
+
+#### Controllers
+- ConfirmationController:
+  - generateConfirmation
+  - sendConfirmationEmail
+  - generateReceipt
+  - updateApplicationStatus
+  - getConfirmationDetails
+  - resendConfirmation
+  - validateConfirmation
+  - trackConfirmationStatus
+
+#### Services
+- ConfirmationService:
+  - confirmationGeneration
+  - emailNotification
+  - receiptGeneration
+  - statusUpdates
+  - auditLogging
+  - nextStepsGeneration
+  - confirmationValidation
+  - communicationTracking
+
+#### Routes
+- POST /api/confirmations/generate
+- GET /api/confirmations/:id
+- POST /api/confirmations/:id/resend
+- GET /api/confirmations/:id/receipt
+- GET /api/confirmations/:id/status
+- POST /api/confirmations/:id/validate
+- GET /api/confirmations/history
+- POST /api/confirmations/:id/next-steps
+
+### Frontend Changes
+#### Components
+- ConfirmationDisplay
+  - SuccessMessage
+  - TransactionDetails
+  - ReceiptDownload
+  - NextStepsGuide
+  - ShareOptions
+- ConfirmationManagement
+  - ConfirmationList
+  - DetailsViewer
+  - ResendTools
+  - StatusTracker
+- ReceiptGenerator
+  - ReceiptTemplate
+  - DownloadOptions
+  - PrintView
+  - SharingTools
+
+#### Services
+- confirmationService:
+  - confirmationHandling
+  - receiptManagement
+  - notificationDelivery
+  - statusTracking
+  - historyManagement
+
+### Dependencies
+- Authentication System ✅ (#1)
+- Stripe Integration ✅ (#14)
+- Payment Processing ✅ (#15)
+
+### Special Features
+1. Confirmation Display
+   - Interactive success animation
+   - Responsive design
+   - Printer-friendly version
+   - Social sharing options
+   - QR code generation
+
+2. Communication Features
+   - Multi-channel notifications
+   - Template customization
+   - Language localization
+   - Delivery tracking
+   - Retry mechanisms
+
+3. Receipt Features
+   - Multiple formats (PDF, HTML)
+   - Digital signature
+   - Archival system
+   - Batch processing
+   - Custom branding
+
+4. Next Steps Guidance
+   - Dynamic instructions
+   - Timeline visualization
+   - Requirement checklist
+   - Progress tracking
+   - Reminder system
+
+### Implementation Notes
+1. Confirmation Process
+   - Define confirmation flow
+   - Set up templates
+   - Configure notifications
+   - Implement tracking
+   - Set up archiving
+
+2. Communication Setup
+   - Email configuration
+   - Template system
+   - Notification rules
+   - Delivery tracking
+   - Error handling
+
+3. Receipt Generation
+   - Template design
+   - PDF generation
+   - Digital signing
+   - Storage setup
+   - Access control
+
+4. Testing Strategy
+   - Confirmation flow
+   - Email delivery
+   - Receipt generation
+   - Status updates
+   - Integration testing
