+++ .github/ISSUES/2-user-profile-management.md
@@ -2,17 +2,17 @@
 
 ## Requirements
 ### User Story
-"As a user, I want to manage my profile information so that I can keep my account details up to date"
+"As a user, I want to manage my profile information so that I can easily apply for visas and track my applications"
 
 ### Acceptance Criteria
 - [ ] Users can view their profile details
 - [ ] Users can update their:
-  - Username
-  - Email
-  - Password
-  - Profile picture (optional)
-  - Personal information
-- [ ] Users can delete their account
+  - Personal Information
+  - Contact Details
+  - Passport Information
+  - Communication preferences
+- [ ] Users can view their application history
+- [ ] Users can track application status
 - [ ] All changes require authentication
 - [ ] Email changes require verification
 
@@ -22,51 +22,66 @@
 #### Models
 - Update User Model:  ```javascript
   {
-    profile_image: String,
-    phone: String,
-    address: {
-      street: String,
-      city: String,
-      state: String,
-      zipCode: String,
-      country: String
+    personal_info: {
+      full_name: String,
+      date_of_birth: Date,
+      nationality: String,
+      passport_number: String,
+      passport_expiry: Date
+    },
+    contact_info: {
+      email: String,
+      phone: String,
+      address: {
+        street: String,
+        city: String,
+        state: String,
+        zipCode: String,
+        country: String
+      }
     },
     preferences: {
       notifications: Boolean,
-      newsletter: Boolean
-    }
+      language: String
+    },
+    applications: [{
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'Application'
+    }]
   }  ```
 
 #### Controllers
 - ProfileController:
   - getProfile
   - updateProfile
-  - updatePassword
-  - deleteAccount
-  - updateEmail
+  - updatePassportInfo
+  - getApplicationHistory
+  - updateContactInfo
 
 #### Routes
 - GET /api/profile
 - PUT /api/profile
-- PUT /api/profile/password
-- DELETE /api/profile
-- PUT /api/profile/email
+- GET /api/profile/applications
+- PUT /api/profile/passport
+- PUT /api/profile/contact
 
 ### Frontend Changes
 #### Components
 - ProfileLayout
 - ProfileForm
-- PasswordChangeForm
-- DeleteAccountModal
-- ImageUploader
+- PassportInfoForm
+- ApplicationHistoryList
+- ContactInfoForm
+- StatusTracker
 
 #### Services
 - profileService:
   - Profile CRUD operations
-  - Image upload
-  - Email change verification
+  - Application history retrieval
+  - Status tracking
+  - Document management
 
 ### Dependencies
-- Authentication System ✅ (#1)
-- File upload service (for profile images)
-- Email service (for verification) 
+- Authentication System (#1)
+- Document Upload System (#10)
+- Application Status System
