+++ .github/ISSUES/3-admin-dashboard.md
@@ -2,20 +2,36 @@
 
 ## Requirements
 ### User Story
-"As an admin, I want to manage users and monitor system activity so that I can maintain the platform effectively"
+"As an admin, I want to manage visa applications and users so that I can process visa requests efficiently"
 
 ### Acceptance Criteria
-- [ ] Admin can view all users
-- [ ] Admin can manage user roles
-- [ ] Admin can disable/enable user accounts
-- [ ] Admin can view system metrics
-- [ ] Admin can view audit logs
+- [ ] Admin can view all visa applications
+- [ ] Admin can update application status (pending, approved, rejected)
+- [ ] Admin can view and verify uploaded documents
+- [ ] Admin can manage user accounts
+- [ ] Admin can view payment status
+- [ ] Admin can generate basic reports
 - [ ] Access restricted to admin role only
 
 ## Technical Design
 
 ### Backend Changes
 #### Models
+- Update Application Model:  ```javascript
+  {
+    status: {
+      type: String,
+      enum: ['pending', 'reviewing', 'approved', 'rejected'],
+      default: 'pending'
+    },
+    adminNotes: String,
+    reviewedBy: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'User'
+    },
+    reviewedAt: Date
+  }  ```
+
 - Update User Model:  ```javascript
   {
     role: {
@@ -30,46 +46,43 @@
     }
   }  ```
 
-- New AuditLog Model:  ```javascript
-  {
-    action: String,
-    userId: ObjectId,
-    targetId: ObjectId,
-    details: Object,
-    timestamp: Date
-  }  ```
-
 #### Controllers
 - AdminController:
+  - getApplicationsList
+  - updateApplicationStatus
   - getUsersList
-  - updateUserRole
-  - updateUserStatus
-  - getSystemMetrics
-  - getAuditLogs
+  - getApplicationStats
+  - verifyDocuments
+  - getPaymentStatus
 
 #### Routes
+- GET /api/admin/applications
+- PUT /api/admin/applications/:id/status
 - GET /api/admin/users
-- PUT /api/admin/users/:id/role
-- PUT /api/admin/users/:id/status
-- GET /api/admin/metrics
-- GET /api/admin/logs
+- GET /api/admin/stats
+- PUT /api/admin/documents/:id/verify
+- GET /api/admin/payments
 
 ### Frontend Changes
 #### Components
-- AdminLayout
-- UserManagementTable
-- SystemMetricsDisplay
-- AuditLogViewer
-- RoleManager
-- AdminDashboardStats
+- AdminDashboard
+  - ApplicationsTable
+  - StatusUpdateModal
+  - DocumentVerificationView
+  - UserManagementTable
+  - PaymentStatusView
+  - StatsDisplay
 
 #### Services
 - adminService:
-  - User management operations
-  - System metrics fetching
-  - Audit log retrieval
+  - Application management
+  - Document verification
+  - User management
+  - Stats compilation
+  - Payment tracking
 
 ### Dependencies
 - Authentication System ✅ (#1)
-- User Profile Management (#2)
-- Admin role implementation 
+- User Profile Management ✅ (#2)
+- Document Upload System (#10)
+- Payment System (#13)
