--- .github/ISSUES/6-product-search-filter.md
@@ -1,125 +1,0 @@
-# Product Search & Filter
-
-## Requirements
-### User Story
-"As a customer, I want to search and filter products so that I can easily find items I'm interested in"
-
-### Acceptance Criteria
-- [ ] Users can search products by name and description
-- [ ] Users can filter products by:
-  - Category
-  - Price range
-  - Status (in stock/out of stock)
-  - Rating
-- [ ] Users can sort products by:
-  - Price (low to high/high to low)
-  - Name (A-Z/Z-A)
-  - Latest
-  - Popularity
-- [ ] Search results update in real-time
-- [ ] Mobile-responsive filter interface
-
-## Technical Design
-
-### Backend Changes
-#### Models
-- Update Product Model:  ```javascript
-  {
-    searchIndex: { 
-      type: String,
-      index: true 
-    },
-    popularity_score: {
-      type: Number,
-      default: 0,
-      index: true
-    },
-    meta: {
-      views: { type: Number, default: 0 },
-      searches: { type: Number, default: 0 },
-      last_viewed: Date
-    }
-  }  ```
-
-#### Controllers
-- SearchController:
-  - searchProducts
-    - Full-text search
-    - Filter application
-    - Sort handling
-    - Pagination
-  - getFilters
-    - Available categories
-    - Price ranges
-    - Dynamic filters
-  - updateSearchMetrics
-    - Track popular searches
-    - Update product scores
-
-#### Routes
-- GET /api/products/search
-- GET /api/products/filters
-- GET /api/products/suggestions
-- POST /api/products/search-metrics
-
-#### Middleware
-- validateSearchParams
-- handleSearchTracking
-- cacheSearchResults
-- rateLimitSearch
-
-### Frontend Changes
-#### Components
-- SearchBar
-  - Autocomplete
-  - Search history
-  - Suggestions
-- FilterPanel
-  - Price range slider
-  - Category checkboxes
-  - Status toggles
-  - Mobile drawer
-- SortingControls
-  - Sort options
-  - View options (grid/list)
-- SearchResults
-  - Result count
-  - Loading states
-  - No results handling
-
-#### Services
-- searchService:
-  - Search operations
-  - Filter management
-  - Search history
-  - Analytics tracking
-
-### Dependencies
-- Product CRUD (#4)
-- Category Management (#5)
-- Redis Cache Service
-- Search Index Service
-
-### Testing Requirements
-- Unit Tests:
-  - Search algorithm
-  - Filter combinations
-  - Sort functions
-- Integration Tests:
-  - Search API
-  - Filter application
-  - Sort operations
-- E2E Tests:
-  - Search flow
-  - Filter interactions
-  - Mobile responsiveness
-
-### Documentation Updates
-- API Documentation:
-  - Search parameters
-  - Filter options
-  - Response format
-- User Guide:
-  - Search tips
-  - Filter usage
-  - Sort options
