--- .github/ISSUES/9-checkout-process.md
@@ -1,188 +1,0 @@
-# Checkout Process
-
-## Requirements
-### User Story
-"As a customer, I want to complete my purchase by providing shipping and payment information"
-
-### Acceptance Criteria
-- [ ] Users can:
-  - Enter shipping information
-  - Select shipping method
-  - Review order summary
-  - Enter payment information
-  - Place order
-- [ ] System should:
-  - Validate shipping address
-  - Calculate shipping costs
-  - Calculate tax
-  - Verify stock availability
-  - Process payment securely
-  - Generate order confirmation
-- [ ] Support guest checkout
-- [ ] Save shipping info for registered users
-
-## Technical Design
-
-### Backend Changes
-#### Models
-- Order Model:
-  ```javascript
-  {
-    user: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'User',
-      default: null
-    },
-    guest_email: {
-      type: String,
-      required: function() { return !this.user; }
-    },
-    items: [{
-      product: {
-        type: mongoose.Schema.Types.ObjectId,
-        ref: 'Product',
-        required: true
-      },
-      quantity: {
-        type: Number,
-        required: true,
-        min: 1
-      },
-      price: {
-        type: Number,
-        required: true
-      }
-    }],
-    shipping: {
-      address: {
-        street: String,
-        city: String,
-        state: String,
-        zip: String,
-        country: String
-      },
-      method: {
-        type: String,
-        required: true
-      },
-      cost: {
-        type: Number,
-        required: true
-      }
-    },
-    payment: {
-      method: String,
-      transaction_id: String,
-      status: {
-        type: String,
-        enum: ['pending', 'completed', 'failed'],
-        default: 'pending'
-      }
-    },
-    subtotal: Number,
-    tax: Number,
-    total: Number,
-    status: {
-      type: String,
-      enum: ['pending', 'processing', 'shipped', 'delivered', 'cancelled'],
-      default: 'pending'
-    },
-    created_at: {
-      type: Date,
-      default: Date.now
-    }
-  }
-  ```
-
-#### Controllers
-- CheckoutController:
-  - initializeCheckout
-    - Validate cart
-    - Check stock
-    - Calculate initial totals
-  - validateShipping
-    - Address validation
-    - Shipping cost calculation
-  - processPayment
-    - Payment gateway integration
-    - Transaction handling
-  - createOrder
-    - Convert cart to order
-    - Update inventory
-    - Clear cart
-  - getOrderStatus
-    - Order tracking
-    - Status updates
-
-#### Routes
-- POST /api/checkout/initialize
-- POST /api/checkout/shipping
-- POST /api/checkout/payment
-- POST /api/checkout/complete
-- GET /api/checkout/status/:id
-
-#### Middleware
-- validateCheckoutSession
-- checkStockAvailability
-- validateShippingAddress
-- handlePaymentSecurity
-- ensureCartNotEmpty
-
-### Frontend Changes
-#### Components
-- CheckoutFlow
-  - Progress indicator
-  - Step navigation
-  - Order summary
-- ShippingForm
-  - Address input
-  - Method selection
-  - Validation feedback
-- PaymentForm
-  - Secure payment input
-  - Payment method selection
-- OrderReview
-  - Order details
-  - Final confirmation
-- OrderConfirmation
-  - Success message
-  - Order details
-  - Next steps
-
-#### Services
-- checkoutService:
-  - Checkout flow management
-  - Shipping calculations
-  - Payment processing
-  - Order creation
-
-### Dependencies
-- Authentication System (#1)
-- Product CRUD (#4)
-- Cart Management (#8)
-- Payment Gateway Service
-- Shipping API Integration
-
-### Testing Requirements
-- Unit Tests:
-  - Price calculations
-  - Address validation
-  - Stock verification
-- Integration Tests:
-  - Checkout flow
-  - Payment processing
-  - Order creation
-- E2E Tests:
-  - Complete checkout flow
-  - Error handling
-  - Order confirmation
-
-### Documentation Updates
-- API Documentation:
-  - Checkout process
-  - Payment integration
-  - Order structure
-- User Guide:
-  - Checkout steps
-  - Payment options
-  - Order tracking
