+++ PRD.md
@@ -1,147 +1,109 @@
 # Product Requirements Document
 
 ## Project Overview
-An e-commerce platform built with MERN stack enabling users to browse, purchase, and manage products with secure payment processing.
+A serverless visa application platform built with Next.js enabling users to apply for tourist visas with a simple, guided process and status tracking.
 
 ### Project Goals
-1. Create a user-friendly e-commerce platform
-2. Implement secure user authentication
-3. Provide efficient product management
+1. Create user-friendly visa application platform
+2. Implement secure document handling
+3. Provide efficient application tracking
 4. Enable secure payment processing
-5. Deliver smooth order management
+5. Deliver simple admin management
 
 ## Detailed Requirements
 
 ### 1. User Management System
-#### Customer Features
-- Registration with email verification
-- Secure login/logout
-- Password reset functionality
+#### User Features
+- Basic registration/login
 - Profile management
   - Personal information
-  - Address management
-  - Order history
-  - Wishlist
+  - Application history
+  - Status tracking
+- Document upload/management
 
 #### Admin Features
-- User management dashboard
-- User role management
-- Activity monitoring
-- Access control management
+- Simple application dashboard
+- Basic status management
+- Application review interface
 
-### 2. Product Management System
-#### Catalog Features
-- Product categories and subcategories
-- Advanced search functionality
-  - By name
-  - By category
-  - By price range
-  - By availability
-- Filtering and sorting options
-- Product reviews and ratings
+### 2. Visa Application System
+#### Application Features
+- 7 Tourist Visa Forms:
+  - Costa Rica
+  - USA
+  - India
+  - Reino Unido
+  - Cuba
+  - Tailandia
+  - Egipto
+- Each form includes:
+  - Personal information
+  - Travel details
+  - Document uploads
+  - Payment processing
 
-#### Product Details
-- Product name and description
-- Price and discount information
-- Stock availability
-- Product images (multiple)
-- Product variations (size, color)
-- Related products
-- Customer reviews section
-
-### 3. Shopping Experience
-#### Cart Management
-- Add/remove items
-- Update quantities
-- Save for later
-- Price calculations
-  - Subtotal
-  - Tax
-  - Shipping
-  - Total
-
-#### Checkout Process
-- Address selection/input
-- Shipping method selection
-- Payment method selection
-- Order review
-- Order confirmation
+#### Application Process
+- Step-by-step form completion
+- Document upload capability
+- Payment integration
+- Status tracking
 - Email notifications
 
-### 4. Payment System
+### 3. Payment System
 #### Stripe Integration
 - Secure payment processing
-- Multiple payment methods
-  - Credit/debit cards
-  - Digital wallets
-- Payment verification
-- Refund handling
+- Basic payment verification
+- Payment confirmation
 
-### 5. Order Management
-#### Customer Features
-- Order tracking
-- Order history
-- Order status updates
-- Order cancellation
-- Return requests
-
-#### Admin Features
-- Order processing
-- Status management
-- Return handling
-- Refund processing
+### 4. Admin Management
+#### Basic Features
+- Application review dashboard
+- Status update capability
+- Basic application management
 
 ## Technical Requirements
 
 ### 1. Security
-- JWT authentication
-- Password encryption
-- CORS configuration
-- Rate limiting
+- Next.js authentication
+- Document storage security
 - Input validation
-- XSS protection
-- CSRF protection
+- Rate limiting
+- Basic CORS setup
 
 ### 2. Performance
-- Page load < 3 seconds
-- Mobile responsiveness
-- Image optimization
-- Efficient database queries
-- Caching implementation
+- Quick form loading
+- Responsive design
+- Efficient file handling
 - Error handling
 - Loading states
 
 ### 3. Database Design
-- Efficient schema design
-- Proper indexing
-- Data validation
-- Relationship management
-- Backup strategy
+- User profiles
+- Application storage
+- Document references
+- Status tracking
+- Payment records
 
 ### 4. API Design
-- RESTful architecture
-- Proper error handling
-- Request validation
+- Next.js API routes
+- Serverless functions
+- Error handling
 - Response formatting
-- API documentation
 
 ## Success Metrics
 1. User Experience
-   - Successful order rate
-   - Cart abandonment rate
-   - Search success rate
+   - Form completion rate
+   - Application success rate
    - Page load times
 
 2. Technical Performance
    - API response times
    - Error rates
-   - System uptime
-   - Database performance
+   - Upload success rates
 
 3. Business Goals
-   - User registration rate
-   - Order completion rate
-   - Customer return rate
+   - Application completion rate
+   - Payment success rate
    - Admin efficiency metrics
 
 ## Future Considerations
