--- src/tests/esta-test-data.js
@@ -1,277 +1,0 @@
-export const testUsers = [
-  {
-    // User 1: Standard case with all fields
-    fullName: 'Juan Carlos Rodríguez Pérez',
-    birthDate: '1985-06-15',
-    birthCity: 'Madrid',
-    birthCountry: 'España',
-    nationality: 'Española',
-    email: 'juan.rodriguez@email.com',
-    phone: '+34 612 345 678',
-    address: 'Calle Mayor 123, Piso 4B\nMadrid, 28001\nEspaña',
-    fatherName: 'Antonio Rodríguez García',
-    motherName: 'María Pérez Sánchez',
-    passportNumber: 'PAB123456',
-    passportIssueDate: '2020-03-10',
-    passportExpiryDate: '2030-03-09',
-    passportIssuingCountry: 'España',
-    previousUSTravel: true,
-    usAddress: 'Hilton Times Square\n234 West 42nd Street\nNew York, NY 10036',
-    employer: 'Tecnología Innovadora S.L.',
-    jobTitle: 'Ingeniero de Software Senior',
-    workAddress: 'Paseo de la Castellana 95\nMadrid, 28046\nEspaña',
-    hasCriminalRecord: false,
-    termsAccepted: true
-  },
-  {
-    // User 2: Young student
-    fullName: 'Ana María López Silva',
-    birthDate: '2000-09-23',
-    birthCity: 'Barcelona',
-    birthCountry: 'España',
-    nationality: 'Española',
-    email: 'ana.lopez@estudiante.edu',
-    phone: '+34 623 456 789',
-    address: 'Avenida Diagonal 456\nBarcelona, 08006\nEspaña',
-    fatherName: 'José López Martín',
-    motherName: 'Carmen Silva Ruiz',
-    passportNumber: 'PAC234567',
-    passportIssueDate: '2022-01-15',
-    passportExpiryDate: '2032-01-14',
-    passportIssuingCountry: 'España',
-    previousUSTravel: false,
-    usAddress: 'University Residence Hall\n500 College Avenue\nBoston, MA 02215',
-    employer: 'Universidad de Barcelona',
-    jobTitle: 'Estudiante',
-    workAddress: 'Gran Via de les Corts Catalanes 585\nBarcelona, 08007\nEspaña',
-    hasCriminalRecord: false,
-    termsAccepted: true
-  },
-  {
-    // User 3: Business executive
-    fullName: 'Miguel Ángel Fernández Torres',
-    birthDate: '1975-12-03',
-    birthCity: 'Valencia',
-    birthCountry: 'España',
-    nationality: 'Española',
-    email: 'miguel.fernandez@empresa.com',
-    phone: '+34 634 567 890',
-    address: 'Calle Colón 78\nValencia, 46004\nEspaña',
-    fatherName: 'Roberto Fernández Navarro',
-    motherName: 'Isabel Torres Gómez',
-    passportNumber: 'PAD345678',
-    passportIssueDate: '2021-07-20',
-    passportExpiryDate: '2031-07-19',
-    passportIssuingCountry: 'España',
-    previousUSTravel: true,
-    usAddress: 'Marriott Marquis\n1535 Broadway\nNew York, NY 10036',
-    employer: 'Corporación Global S.A.',
-    jobTitle: 'Director Ejecutivo',
-    workAddress: 'Paseo de la Alameda 48\nValencia, 46023\nEspaña',
-    hasCriminalRecord: false,
-    termsAccepted: true
-  },
-  {
-    // User 4: Retired person
-    fullName: 'María del Carmen Ruiz Ortega',
-    birthDate: '1955-03-28',
-    birthCity: 'Sevilla',
-    birthCountry: 'España',
-    nationality: 'Española',
-    email: 'mcarmen.ruiz@email.com',
-    phone: '+34 645 678 901',
-    address: 'Avenida de la Constitución 23\nSevilla, 41004\nEspaña',
-    fatherName: 'Francisco Ruiz Moreno',
-    motherName: 'Dolores Ortega Jiménez',
-    passportNumber: 'PAE456789',
-    passportIssueDate: '2019-11-05',
-    passportExpiryDate: '2029-11-04',
-    passportIssuingCountry: 'España',
-    previousUSTravel: false,
-    usAddress: 'Holiday Inn Miami Beach\n4333 Collins Avenue\nMiami Beach, FL 33140',
-    employer: 'Jubilada',
-    jobTitle: 'Jubilada',
-    workAddress: 'N/A',
-    hasCriminalRecord: false,
-    termsAccepted: true
-  },
-  {
-    // User 5: Artist
-    fullName: 'Pablo Martín Sánchez',
-    birthDate: '1990-08-12',
-    birthCity: 'Málaga',
-    birthCountry: 'España',
-    nationality: 'Española',
-    email: 'pablo.martin@artista.com',
-    phone: '+34 656 789 012',
-    address: 'Calle Larios 45\nMálaga, 29005\nEspaña',
-    fatherName: 'Diego Martín Vega',
-    motherName: 'Laura Sánchez Castro',
-    passportNumber: 'PAF567890',
-    passportIssueDate: '2023-02-28',
-    passportExpiryDate: '2033-02-27',
-    passportIssuingCountry: 'España',
-    previousUSTravel: true,
-    usAddress: 'Chelsea Hotel\n222 West 23rd Street\nNew York, NY 10011',
-    employer: 'Freelance',
-    jobTitle: 'Artista Visual',
-    workAddress: 'Estudio de Arte\nCalle de la Victoria 15\nMálaga, 29012\nEspaña',
-    hasCriminalRecord: false,
-    termsAccepted: true
-  },
-  {
-    // User 6: Medical professional
-    fullName: 'Laura García Mendoza',
-    birthDate: '1982-04-17',
-    birthCity: 'Bilbao',
-    birthCountry: 'España',
-    nationality: 'Española',
-    email: 'laura.garcia@hospital.es',
-    phone: '+34 667 890 123',
-    address: 'Gran Vía 89\nBilbao, 48011\nEspaña',
-    fatherName: 'Manuel García López',
-    motherName: 'Ana Mendoza Ruiz',
-    passportNumber: 'PAG678901',
-    passportIssueDate: '2022-09-15',
-    passportExpiryDate: '2032-09-14',
-    passportIssuingCountry: 'España',
-    previousUSTravel: true,
-    usAddress: 'Medical Conference Center\n123 Healthcare Avenue\nChicago, IL 60601',
-    employer: 'Hospital Universitario',
-    jobTitle: 'Médico Especialista',
-    workAddress: 'Alameda Urquijo 35\nBilbao, 48008\nEspaña',
-    hasCriminalRecord: false,
-    termsAccepted: true
-  },
-  {
-    // User 7: Professional athlete
-    fullName: 'Carlos Herrera Blanco',
-    birthDate: '1995-11-30',
-    birthCity: 'Zaragoza',
-    birthCountry: 'España',
-    nationality: 'Española',
-    email: 'carlos.herrera@deporte.es',
-    phone: '+34 678 901 234',
-    address: 'Paseo Independencia 67\nZaragoza, 50004\nEspaña',
-    fatherName: 'Javier Herrera Martínez',
-    motherName: 'Patricia Blanco Sanz',
-    passportNumber: 'PAH789012',
-    passportIssueDate: '2021-04-30',
-    passportExpiryDate: '2031-04-29',
-    passportIssuingCountry: 'España',
-    previousUSTravel: true,
-    usAddress: 'Athletic Complex\n789 Sports Boulevard\nLos Angeles, CA 90015',
-    employer: 'Club Deportivo Professional',
-    jobTitle: 'Deportista Profesional',
-    workAddress: 'Centro Deportivo\nAvenida del Deporte 45\nZaragoza, 50009\nEspaña',
-    hasCriminalRecord: false,
-    termsAccepted: true
-  },
-  {
-    // User 8: Teacher
-    fullName: 'Isabel Moreno Castro',
-    birthDate: '1988-07-25',
-    birthCity: 'Murcia',
-    birthCountry: 'España',
-    nationality: 'Española',
-    email: 'isabel.moreno@colegio.edu',
-    phone: '+34 689 012 345',
-    address: 'Avenida de la Libertad 34\nMurcia, 30009\nEspaña',
-    fatherName: 'Andrés Moreno Pérez',
-    motherName: 'Rosa Castro García',
-    passportNumber: 'PAI890123',
-    passportIssueDate: '2020-08-12',
-    passportExpiryDate: '2030-08-11',
-    passportIssuingCountry: 'España',
-    previousUSTravel: false,
-    usAddress: 'Educational Conference Center\n456 Learning Way\nBoston, MA 02108',
-    employer: 'Colegio San José',
-    jobTitle: 'Profesora de Secundaria',
-    workAddress: 'Calle de la Enseñanza 12\nMurcia, 30005\nEspaña',
-    hasCriminalRecord: false,
-    termsAccepted: true
-  },
-  {
-    // User 9: Entrepreneur
-    fullName: 'Alejandro Torres Navarro',
-    birthDate: '1980-02-14',
-    birthCity: 'Alicante',
-    birthCountry: 'España',
-    nationality: 'Española',
-    email: 'alejandro.torres@startup.com',
-    phone: '+34 690 123 456',
-    address: 'Rambla Méndez Núñez 28\nAlicante, 03002\nEspaña',
-    fatherName: 'Miguel Torres Ruiz',
-    motherName: 'Carmen Navarro López',
-    passportNumber: 'PAJ901234',
-    passportIssueDate: '2023-01-20',
-    passportExpiryDate: '2033-01-19',
-    passportIssuingCountry: 'España',
-    previousUSTravel: true,
-    usAddress: 'Silicon Valley Convention Center\n345 Innovation Drive\nSan Francisco, CA 94105',
-    employer: 'TechStart S.L.',
-    jobTitle: 'CEO y Fundador',
-    workAddress: 'Parque Tecnológico\nAvenida Innovación 56\nAlicante, 03008\nEspaña',
-    hasCriminalRecord: false,
-    termsAccepted: true
-  },
-  {
-    // User 10: Journalist
-    fullName: 'Elena Jiménez Romero',
-    birthDate: '1992-05-08',
-    birthCity: 'Granada',
-    birthCountry: 'España',
-    nationality: 'Española',
-    email: 'elena.jimenez@noticias.es',
-    phone: '+34 601 234 567',
-    address: 'Calle Recogidas 45\nGranada, 18005\nEspaña',
-    fatherName: 'Fernando Jiménez Díaz',
-    motherName: 'Marta Romero Sánchez',
-    passportNumber: 'PAK012345',
-    passportIssueDate: '2021-12-05',
-    passportExpiryDate: '2031-12-04',
-    passportIssuingCountry: 'España',
-    previousUSTravel: true,
-    usAddress: 'Press Center\n567 Media Avenue\nWashington, DC 20001',
-    employer: 'Diario Nacional',
-    jobTitle: 'Periodista Internacional',
-    workAddress: 'Centro de Prensa\nCalle Periodista 23\nGranada, 18014\nEspaña',
-    hasCriminalRecord: false,
-    termsAccepted: true
-  }
-];
-
-// Function to fill form with test data
-export const fillFormWithTestData = (userData) => {
-  // This function can be used to automatically fill the form
-  Object.keys(userData).forEach(field => {
-    const input = document.querySelector(`[name="${field}"]`);
-    if (input) {
-      if (input.type === 'checkbox') {
-        input.checked = userData[field];
-      } else {
-        input.value = userData[field];
-      }
-      // Trigger change event to update form state
-      input.dispatchEvent(new Event('change', { bubbles: true }));
-    }
-  });
-};
-
-// Example usage in tests:
-/*
-import { testUsers, fillFormWithTestData } from './esta-test-data';
-
-describe('ESTA Form Tests', () => {
-  test('should fill form with test user data', () => {
-    testUsers.forEach((userData, index) => {
-      // Clear form
-      // Fill form with test data
-      fillFormWithTestData(userData);
-      // Verify form values
-      // Submit form
-      // Verify submission
-    });
-  });
-});
-*/
