+++ version-1.0.0.md
@@ -1,222 +1,117 @@
-# Version 1.0.0 - Basic E-commerce Platform
+# Version 1.0.0 - Visa Application Platform
 
-This document outlines the implementation plan for the initial release of the E-commerce MERN application.
-
-## Backend Development
-
-### 1. Basic API Setup
-- Express Configuration
-  - Server setup
-  - Basic middleware (cors, helmet)
-  - Error handling middleware
-  - Environment configuration
-
-- Route Structure
-  - Base router setup
-  - Controller structure
-  - Basic response formatting
-
-### 2. Database Models
-- User Model
-  - Username
-  - Email
-  - Password (hashed)
-  - Role (customer/admin)
-  - Timestamps
-
-- Product Model
-  - Name
-  - Description
-  - Price
-  - Stock quantity
-  - Image URL
-  - Category
-  - Timestamps
-
-- Cart Model
-  - UserId
-  - Items (array of products with quantities)
-  - Total price
-  - Timestamps
-
-- Order Model
-  - UserId
-  - Items purchased
-  - Total amount
-  - Shipping address
-  - Order status (pending, completed)
-  - Payment status
-  - Timestamps
-
-### 3. Core API Features
-- Basic CRUD Operations
-  - Controller implementation
-  - Route handlers
-  - Input validation
-  - Error responses
-
-- Testing Structure
-  - Test environment setup
-  - Basic test patterns
-  - API endpoint tests
-
-### 4. Authentication System
-- [x] User Registration
-- [x] Email Verification
-- [x] User Login
-- [x] Password Reset
-  - [x] Forgot Password Request
-  - [x] Reset Password with Token
-- [ ] User Profile Management (next feature)
-
-### Completed Authentication Features
-1. Registration with email verification
-2. Login with JWT authentication
-3. Password reset flow
-   - Request reset link
-   - Email delivery
-   - Token verification
-   - Password update
-
-### 5. Feature Endpoints
-- Product Management
-  - List products
-  - Single product details
-  - Admin product operations
-
-- Cart Operations
-  - Get user's cart
-  - Add/update items
-  - Remove items
-  - Calculate totals
-
-- Order Processing
-  - Create orders
-  - Order history
-  - Order status
-
-### 6. Payment Integration
-- Stripe Setup
-  - Basic configuration
-  - Payment intent creation
-  - Success/failure handling
+This document outlines the implementation plan for the initial release of the Next.js serverless visa application platform.
 
 ## Frontend Development
 
 ### 1. Base Setup
-- React Configuration
+- Next.js Configuration
   - Project structure
   - Routing setup
   - State management
   - API service setup
+  - Authentication setup
 
-### 2. Authentication Pages
-- Login Page
-  - Email/password form
-  - Basic validation
+### 2. User Authentication
+- Login/Register Pages
+  - Basic forms
+  - Validation
   - Error handling
+  - Success states
 
-- Registration Page
-  - Registration form
-  - Basic validation
-  - Success/error messages
+### 3. Visa Application Forms
+- 7 Tourist Visa Forms:
+  - Costa Rica Form
+  - USA Form
+  - India Form
+  - Reino Unido Form
+  - Cuba Form
+  - Tailandia Form
+  - Egipto Form
 
-### 3. Product Features
-- Product List
-  - Grid display
-  - Basic search
+- Common Features for Each:
+  - Multi-step form process
+  - Document upload
+  - Form validation
+  - Progress saving
+  - Status display
+
+### 4. User Dashboard
+- Application List
+  - Status overview
+  - Application details
+  - Document management
+
+### 5. Admin Dashboard
+- Application Management
+  - Application list view
+  - Status updates
+  - Document review
   - Basic filtering
 
-- Product Detail
-  - Product information
-  - Add to cart button
-  - Stock status
+### 6. Payment Integration
+- Stripe Setup
+  - Payment flow
+  - Success/failure handling
+  - Payment confirmation
 
-### 4. Shopping Cart
-- Cart Management
-  - Add/remove items
-  - Update quantities
-  - Price calculations
+## API Routes Development
 
-- Cart Summary
-  - Item list
-  - Total price
+### 1. Authentication Routes
+- User registration
+- Login
+- Profile management
 
-### 5. Checkout Process
-- Checkout Flow
-  - Shipping information
-  - Basic Stripe integration
+### 2. Application Routes
+- Form submission
+- Status updates
+- Document handling
+- Application retrieval
 
-- Order Confirmation
-  - Success/failure messages
-  - Order details
+### 3. Admin Routes
+- Application management
+- Status management
+- Admin operations
 
-### 6. User Features
-- User Dashboard
-  - Order history
-  - Order details
+### 4. Payment Routes
+- Stripe integration
+- Payment processing
+- Payment verification
 
-- Admin Dashboard
-  - Product management
-  - Order management
-  - Basic inventory control
+## Database Models
+
+### 1. User Model
+- Basic info
+- Auth details
+- Role (user/admin)
+
+### 2. Application Model
+- Applicant details
+- Visa type
+- Status
+- Documents
+- Payment status
 
 ## Testing
-- Backend Testing
-  - API endpoints
-  - Auth flow
-  - Data operations
-
-- Frontend Testing
-  - Component rendering
-  - User interactions
-  - Integration tests
+- Component Testing
+- API Route Testing
+- Integration Testing
+- Form Validation Testing
 
 ## Documentation
-- API Documentation
-  - Endpoint descriptions
-  - Request/response formats
-
 - Setup Guide
-  - Installation steps
-  - Environment setup
+- API Documentation
+- Deployment Guide
 
 ## Security Implementation
-- Basic rate limiting
-- CORS setup
-- Input validation
-- Error handling
-- Secure headers
+- Authentication
+- Input Validation
+- File Upload Security
+- API Route Protection
 
 ## Success Criteria
-- Users can browse and search products
-- Users can manage their cart
-- Checkout process works with Stripe
-- Orders are tracked and managed
-- Admin can manage products and orders
-- Core functionality is tested
-- Documentation is complete
-
-# Version 1.0.0 Implementation Plan
-
-## Progress Tracking
-- [x] Basic API Setup
-- [x] Basic User Model
-- [ ] Authentication System (In Progress)
-  - [ ] JWT Implementation
-  - [ ] Email Verification
-  - [ ] Password Reset
-  - [ ] Frontend Auth Components
-- [ ] Database Models
-  - [ ] Product Model
-  - [ ] Cart Model
-  - [ ] Order Model
-  - [ ] Core API Features
-  - [ ] Testing Structure
-  - [ ] Feature Endpoints
-  - [ ] Payment Integration
-  - [ ] Frontend Development
-  - [ ] Testing
-  - [ ] Documentation
-  - [ ] Security Implementation
-  - [ ] Success Criteria
+- Users can complete visa applications
+- Admins can manage applications
+- Payment processing works
+- Document upload functions
+- Status tracking works
